Clarify list_add documentation
Originally, the function list_add is said to insert a node at the
beginning of a given list. However, it could also be used to insert a
node after any existing node in the list. Therefore, this commit
modifies its documentation to reflect the actual list manipulation
performed by the function.

Change-Id: I661d7a0ebb876712c8ff75deef84370a51e62f1c